Swap IEEE 754 floating points with integral representation.
The idea is that, in case of fixed precision calculations an integral
representation can, with some limitations, serve the same purposei (Just
consider the represented number equal to the original number multiplied by
10^[number of decimal places]).

+++ src/point_cloud_gen.h
@@ -4,7 +4,7 @@
 #include <stdio.h>
 
 typedef unsigned long cloud_size_t;
-typedef double coord_t;
+typedef long coord_t;
 
 typedef struct {
   coord_t x, y;
